fix: restrict dev auth endpoints to non-prod profiles
